LPS-77311 user-associated-data-web - (unrelated) moves classes to appropriate packages
